+++ packages/stream_chat/lib/src/core/api/channel_api.dart
@@ -375,4 +375,24 @@ class ChannelApi {
     );
     return EmptyResponse.fromJson(response.data);
   }
+
+  /// Updates some of the member data
+  Future<PartialUpdateMemberResponse> updateMemberPartial({
+    required String channelId,
+    required String channelType,
+    Map<String, Object?>? set,
+    List<String>? unset,
+  }) async {
+    final response = await _client.patch(
+      // Note: user_id is not required for client side Apis as it can be fetched
+      // directly from the user token but, for the api path is built with it
+      // so we need to pass it as a placeholder.
+      '${_getChannelUrl(channelId, channelType)}/member/{user_id}',
+      data: {
+        if (set != null) 'set': set,
+        if (unset != null) 'unset': unset,
+      },
+    );
+    return PartialUpdateMemberResponse.fromJson(response.data);
+  }
 }