+++ packages/stream_chat/lib/src/db/chat_persistence_client.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:stream_chat/src/core/api/requests.dart';
 import 'package:stream_chat/src/core/api/sort_order.dart';
 import 'package:stream_chat/src/core/models/attachment_file.dart';
@@ -90,8 +91,15 @@ abstract class ChatPersistenceClient {
       getMessagesByCid(cid, messagePagination: messagePagination),
       getPinnedMessagesByCid(cid, messagePagination: pinnedMessagePagination),
     ]);
+
+    final members = data[0] as List<Member>?;
+    final membership = userId == null
+        ? null
+        : members?.firstWhereOrNull((it) => it.userId == userId);
+
     return ChannelState(
-      members: data[0] as List<Member>?,
+      members: members,
+      membership: membership,
       read: data[1] as List<Read>?,
       channel: data[2] as ChannelModel?,
       messages: data[3] as List<Message>?,