+++ packages/stream_chat_persistence/test/src/mapper/member_mapper_test.dart
@@ -21,6 +21,8 @@ void main() {
       invited: math.Random().nextBool(),
       banned: math.Random().nextBool(),
       shadowBanned: math.Random().nextBool(),
+      pinnedAt: DateTime.now(),
+      archivedAt: DateTime.now(),
       isModerator: math.Random().nextBool(),
       extraData: {'test_extra_data': 'testData'},
     );
@@ -35,6 +37,8 @@ void main() {
     expect(member.invited, entity.invited);
     expect(member.banned, entity.banned);
     expect(member.shadowBanned, entity.shadowBanned);
+    expect(member.pinnedAt, isSameDateAs(entity.pinnedAt));
+    expect(member.archivedAt, isSameDateAs(entity.archivedAt));
     expect(member.isModerator, entity.isModerator);
     expect(member.extraData, entity.extraData);
   });
@@ -52,6 +56,8 @@ void main() {
       invited: math.Random().nextBool(),
       banned: math.Random().nextBool(),
       shadowBanned: math.Random().nextBool(),
+      pinnedAt: DateTime.now(),
+      archivedAt: DateTime.now(),
       isModerator: math.Random().nextBool(),
       extraData: const {'test_extra_data': 'testData'},
     );
@@ -67,6 +73,8 @@ void main() {
     expect(entity.invited, member.invited);
     expect(entity.banned, member.banned);
     expect(entity.shadowBanned, member.shadowBanned);
+    expect(entity.pinnedAt, isSameDateAs(member.pinnedAt));
+    expect(entity.archivedAt, isSameDateAs(member.archivedAt));
     expect(entity.isModerator, member.isModerator);
     expect(entity.extraData, member.extraData);
   });